requirements.txt
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cos-python-sdk-v5==1.9.6
crcmod==1.7
cryptography==3.4.7
dicttoxml==1.7.4
et-xmlfile==1.1.0
idna==2.10
openpyxl==3.0.7
pdfminer.six==20201018
peewee==3.14.4
pycparser==2.20
pycryptodome==3.10.1
PyMySQL==1.0.2
requests==2.25.1
six==1.15.0
sortedcontainers==2.4.0
urllib3==1.26.4
